requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-nmap==0.6.1
requests==2.22.0
six==1.14.0
tabulate==0.8.6
urllib3==1.25.8
wrapt==1.11.2
